Implemented Abstract Class structure, see comment.
The approach is as follows:
We will have multiple implementations of the computation of the De
Casteljau Algorithm. Therefore, in order to obey pythons DRY principle,
we have to think which methods are not effected by an change in internal
calculation.

Since we want to implement parallelism (currently just in the form of
threading) we have to choose a function which calculates all points as
the generic one.

Currently, there are 3 obvious ways to calculate the bezier curve:
- Through the normal De Casteljau Algorithm, this class is now called
BezierCurve
- Through a threaded De Casteljau Algorith, this call is now called
BezierCurveThreaded
- Through Bernstein Polynomials, this will be implemented soon.

My approach is as follows: We make the init_func abstract, which means
that the deriving class needs to give us a function to compute an array
of values. This works for all other functions, since:

- __init__ is the same, because we just set the self.init_func
- init_func is abstract
- _get_or_create_values is the same because it just calls self.func on
all values
- curve is the same since it just uses _get_or_create_values
- intersect is the same since it is just a static method
- min_max_box is the same since it just uses bezier_points, which we
don't alter at all
- collision_check is the same since it doesnt use any attributes
- box_collision_check is the same since it just ueses the boxs, which
are derived from the bezier_points, which we don't alter
- curve_collision_check is the same since curve is the same
- plot is the same since curve is the same
- show_funcs is the same since plot is the same

---

This also has another advantage: In the old implementation we had to
make sure that from every point we generate the values we have to make
sure to generate the mix_max_box() as well. Now we again just have 1
point where they are generated.

---

Why I have changed the THolder: Initially, I didn't want to give the
THolder an array and wanted to leave it as implemented before. This
isn't possible since numpy is native code which doesn't work with sympy.
Therefore, I couldn't use the sy.symbol('t') for np.linspace(0,1,t),
which means that I couldn't generate an numpy array within the
lamdify-function.