Standard compliant code restructuring (see comment)
- We have to name the src folder as the name of our package in order to
properly import it by that name.

- Every (sub-)package needs it's own __init__.py. This should be empty
until needed otherwise. See https://stackoverflow.com/questions/448271/what-is-init-py-for
Maybe in the future we also want to include a lot there in order to give
them the ability to include everything.

- the tests foler has to be within the package, otherwise packing tools
like setuptools will confuse it as the `tests` top level module. (it
also helps the developer using the package to confirm that it works
properly).